double instead of float

+++ DTOS/RouteHistoryDTO.cs
@@ -7,7 +7,7 @@
         public char? Status { get; set; }
         public string? VehicleSpeed { get; set; }
         public long? Epoch { get; set; }
-        public float? Latitude { get; set; }
-        public float? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
     }
 }
