+++ Models/EfModels/CircleGeofence.cs
@@ -12,9 +12,9 @@ public partial class CircleGeofence
     public long? Radius { get; set; }
 
 
-    public float? Latitude { get; set; }
+    public double? Latitude { get; set; }
 
-    public float? Longitude { get; set; }
+    public double? Longitude { get; set; }
 
     public bool IsDeleted { get; set; } = false;
