+++ Models/EfModels/PolygonGeofence.cs
@@ -9,10 +9,10 @@ public partial class PolygonGeofence
 
     public long? GeofenceId { get; set; }
 
-    public float? Latitude { get; set; }
+    public double? Latitude { get; set; }
     public bool IsDeleted { get; set; } = false;
 
-    public float? Longitude { get; set; }
+    public double? Longitude { get; set; }
 
     public virtual Geofence? Geofence { get; set; }
 }
